switches origin of email subscription queries
